feat!: maxTouchPoints doi sang u16 — du lieu cua chinh thu vien khong lot qua kieu cua no
THAY DOI PHA VO. `NavigatorFingerprint::max_touch_points`: Option<u8> -> Option<u16>.
0.2.4 -> 0.3.0.

Mang Bayes kem theo crate khai `256` la mot gia tri co the xay ra cua nut
`maxTouchPoints`:

  possibleValues = 0, 1, 2, 3, 5, 9, 10, 20, 40, 256

`256` khong lot vao u8. Bo doc tra None, va None o day KHONG PHAN BIET DUOC voi
"mang khong co du lieu". Nguoi dung doc ra mot cau sai — rang thu vien khong
biet gi ve ho so nay — trong khi thu vien biet. Khong no, khong canh bao, khong
tra Err; do la ly do no song lau.

Bai kiem da chay o trang thai DO truoc khi sua:

  mang khai 10 gia tri cho maxTouchPoints nhung 1 gia tri khong doc lai duoc:
  ["*STRINGIFIED*256"]

Neu viet sau khi sua thi no chi dang noi u16 chua duoc so nho — dieu khong ai
nghi ngo. Bai kem mot hang doi chung: nut co duoi 5 gia tri thi bai TU DO, de
viec thu gon du lieu khong am tham lam no vo nghia.

Tach `parse_stringified_max_touch_points` khoi `parse_stringified_u8` co y: hai
nut khac nhau co mien gia tri khac nhau, gop chung mot ham nghia la doi mien cua
nut nay se lang le doi mien cua nut kia. `hardwareConcurrency` van dung ban u8.

HAI DIEU KHONG PHAI Y DINH NHUNG DI KEM, khai ra chu khong de chung di ke:

1. Rang buoc noi bo giua ba crate ghim "0.2.0", phai nang len "0.3.0" cung luc,
   khong thi cargo tu choi phan giai.
2. Cargo.lock doi format `version = 3` -> `version = 4`. Khong phai thu toi dinh
   doi; giu lai vi MSRV 1.88 da doi hoi cargo moi hon nguong cua v4 (1.78), va
   cong msrv xac nhan: cargo +1.88 check --locked -> ma thoat 0. Diff lock ngoai
   ba dong phien ban va mot dong format thi KHONG co dependency ngoai nao troi.

THU BAN NAY KHONG SUA, do duoc trong cung luot: `hardwareConcurrency` mang CUNG
LOP LOI va nang hon 150 lan.

  maxTouchPoints        khoi luong > u8 =   0.00%   (256 chi o 1 phan phoi)
  hardwareConcurrency   khoi luong > u8 =  26.36%
    bon gia tri vuot: 384, 448, 512, 640
    127/479 phan phoi co dieu kien mat >50% khoi luong, cao nhat mat 100%

Va no te hon o cho mat thanh `.unwrap_or(4)` — mot con so TRONG HOP LY — chu
khong thanh None. Khong gop vi no dat mot cau hoi ma maxTouchPoints khong dat:
`384` nhan cho mot UA Macintosh la gia tri khong ton tai tren doi, nen noi kieu
de phat no ra co the tao mot dau van tay MOI thay vi sua mot dau van tay sai.
Do la quyet dinh ve DU LIEU, khong phai ve kieu. Item rieng.

Cong: fmt 0, clippy 0, test 0 (49/13/11 passed, 2 ignored), msrv 1.88 -> 0.
CHUA chay tools/fingerprint-parity (cong §4 "tay, khi doi thu vien") vi no doc
app/ ma phien khac dang ghi — chay sau khi ho dong cua so.

Plane: VEIL-699

+++ fingerprint-core/src/types/fingerprint.rs
@@ -157,7 +157,10 @@ pub struct NavigatorFingerprint {
     pub vendor_sub: Option<String>,
     /// `navigator.maxTouchPoints`
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_touch_points: Option<u8>,
+    /// So diem cham toi da. `u16` chu khong `u8`: mang Bayes cua chinh thu
+    /// vien liet `256` trong `possibleValues` cua nut nay, va `u8` lam no
+    /// mat im lang thanh `None` — xem VEIL-699.
+    pub max_touch_points: Option<u16>,
     /// `navigator.product` — always "Gecko".
     #[serde(skip_serializing_if = "Option::is_none")]
     pub product: Option<String>,
